refactor: add data functions to get barbershops and the popular barbershops
