+++ app/_data/get-barbershops-by-title-or-service.ts
@@ -1,0 +1,44 @@
+"use server";
+
+import { db } from "../_lib/prisma";
+
+export interface BarberShopsPageProps {
+  searchParams: Promise<{
+    title?: string;
+    service?: string;
+  }>;
+}
+
+export const getBarbershopsByTitleOrService = async ({
+  searchParams,
+}: BarberShopsPageProps) => {
+  const title = (await searchParams).title;
+  const service = (await searchParams).service;
+
+  return db.barbershop.findMany({
+    where: {
+      OR: [
+        title
+          ? {
+              name: {
+                contains: title,
+                mode: "insensitive",
+              },
+            }
+          : {},
+        service
+          ? {
+              services: {
+                some: {
+                  name: {
+                    contains: service,
+                    mode: "insensitive",
+                  },
+                },
+              },
+            }
+          : {},
+      ],
+    },
+  });
+};
